Fix scrutinizer (#454)
* Simplify scrutinizer config

* Specify tests step for scrutinizer

* Specify tests step for scrutinizer

* Delete spelling fix step

* Delete requirements_test.txt

* Merge app parameters constants in one file

* Delete MainHandler
