+++ .github/workflows/coverage.yml
@@ -24,7 +24,6 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install -r requirements.txt
-        pip install -r requirements_test.txt
         pip install -r requirements_dev.txt
 
     - name: Test with pytest
