+++ CONTRIBUTING.md
@@ -55,7 +55,6 @@ be able to work on TabPy changes:
    python -m pip install --upgrade pip
    pip install -r requirements.txt
    pip install -r requirements_dev.txt
-   pip install -r requirements_test.txt
    ```
 
 ## Tests
