+++ setup.py
@@ -66,27 +66,29 @@ def setup_package():
         # but are listed here because different linux distros use custom
         # python installations.  And users can remove packages at any point
         install_requires=[
-            "backports_abc",
             "cloudpickle",
             "configparser",
-            "decorator",
+            "coverage",
+            "coveralls",
             "docopt",
             "future",
             "genson",
+            "hypothesis",
             "jsonschema",
+            "mock",
             "nltk",
             "numpy",
             "pandas",
             "pyopenssl",
-            "python-dateutil",
+            "pytest",
+            "pytest-cov",
             "requests",
             "scipy",
             "simplejson",
-            "singledispatch",
-            "six",
             "sklearn",
             "textblob",
             "tornado",
+            "twisted",
             "urllib3",
         ],
         entry_points={
@@ -97,20 +99,6 @@ def setup_package():
             ],
         },
         setup_requires=["pytest-runner"],
-        tests_require=[
-            "coverage",
-            "coveralls",
-            "hypothesis",
-            "mock",
-            "nltk",
-            "numpy",
-            "pandas",
-            "pytest",
-            "pytest-cov",
-            "scipy",
-            "sklearn",
-            "textblob",
-        ],
         test_suite="pytest",
     )
