--- tabpy/tabpy_server/app/SettingsParameters.py
@@ -1,17 +1,0 @@
-class SettingsParameters:
-    """
-    Application (TabPyApp) settings names
-    """
-
-    TransferProtocol = "transfer_protocol"
-    Port = "port"
-    ServerVersion = "server_version"
-    UploadDir = "upload_dir"
-    CertificateFile = "certificate_file"
-    KeyFile = "key_file"
-    StateFilePath = "state_file_path"
-    ApiVersions = "versions"
-    LogRequestContext = "log_request_context"
-    StaticPath = "static_path"
-    MaxRequestSizeInMb = "max_request_size_in_mb"
-    EvaluateTimeout = "evaluate_timeout"
