+++ tabpy/tabpy_server/handlers/management_handler.py
@@ -6,8 +6,8 @@ from uuid import uuid4 as random_uuid
 
 from tornado import gen
 
-from tabpy.tabpy_server.app.SettingsParameters import SettingsParameters
-from tabpy.tabpy_server.handlers import MainHandler
+from tabpy.tabpy_server.app.app_parameters import SettingsParameters
+from tabpy.tabpy_server.handlers import BaseHandler
 from tabpy.tabpy_server.handlers.base_handler import STAGING_THREAD
 from tabpy.tabpy_server.management.state import get_query_object_path
 from tabpy.tabpy_server.psws.callbacks import on_state_change
@@ -34,7 +34,7 @@ def copy_from_local(localpath, remotepath, is_dir=False):
         shutil.copy(localpath, remotepath)
 
 
-class ManagementHandler(MainHandler):
+class ManagementHandler(BaseHandler):
     def initialize(self, app):
         super(ManagementHandler, self).initialize(app)
         self.port = self.settings[SettingsParameters.Port]
